refactor(lms): modify comments - (날짜 멱등성)https://github.com/next-step/java-lms/pull/580#discussion_r1818277985 - (변수명 단위 네이밍)https://github.com/next-step/java-lms/pull/580#discussion_r1818278370 - (ImageWidthHeight -> ImageSize)https://github.com/next-step/java-lms/pull/580#discussion_r1818278675 - (ImageSize -> ImageFileSize)https://github.com/next-step/java-lms/pull/580#discussion_r1818279935 - (size 불필요한 래핑 반환)https://github.com/next-step/java-lms/pull/580#discussion_r1818279311
