+++ src/main/java/nextstep/courses/domain/vo/session/CoverImage.java
@@ -1,24 +1,24 @@
 package nextstep.courses.domain.vo.session;
 
-import nextstep.courses.domain.vo.session.image.ImageSize;
+import nextstep.courses.domain.vo.session.image.ImageFileSize;
 import nextstep.courses.domain.vo.session.image.ImageType;
-import nextstep.courses.domain.vo.session.image.ImageWidthHeight;
+import nextstep.courses.domain.vo.session.image.ImageSize;
 
 import java.util.Objects;
 
 public class CoverImage {
-    private final ImageSize imageSize;
+    private final ImageFileSize imageFileSize;
     private final ImageType imageType;
-    private final ImageWidthHeight imageWidthHeight;
+    private final ImageSize imageSize;
 
-    public CoverImage(ImageSize imageSize, ImageType imageType, ImageWidthHeight imageWidthHeight) {
-        this.imageSize = imageSize;
+    public CoverImage(ImageFileSize imageFileSize, ImageType imageType, ImageSize imageSize) {
+        this.imageFileSize = imageFileSize;
         this.imageType = imageType;
-        this.imageWidthHeight = imageWidthHeight;
+        this.imageSize = imageSize;
     }
 
     public CoverImage(int size, String imageTypeText, double width, double height) {
-        this(new ImageSize(size), ImageType.toImageType(imageTypeText), new ImageWidthHeight(width, height));
+        this(new ImageFileSize(size), ImageType.toImageType(imageTypeText), new ImageSize(width, height));
     }
 
     @Override
@@ -26,11 +26,11 @@ public class CoverImage {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         CoverImage that = (CoverImage) o;
-        return Objects.equals(imageSize, that.imageSize) && imageType == that.imageType && Objects.equals(imageWidthHeight, that.imageWidthHeight);
+        return Objects.equals(imageFileSize, that.imageFileSize) && imageType == that.imageType && Objects.equals(imageSize, that.imageSize);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(imageSize, imageType, imageWidthHeight);
+        return Objects.hash(imageFileSize, imageType, imageSize);
     }
 }
