+++ src/main/java/nextstep/courses/domain/vo/session/image/ImageFileSize.java
@@ -1,0 +1,31 @@
+package nextstep.courses.domain.vo.session.image;
+
+import nextstep.courses.ImageSizeExceededException;
+
+import java.util.Objects;
+
+public class ImageFileSize {
+    public static final int MAX_IMAGE_SIZE_MB = 1024 * 1024;
+    public static final String IMAGE_SIZE_EXCEED_MESSAGE = "이미지 크기는 1MB 이하여야 한다";
+    private final int size;
+
+    public ImageFileSize(int size) {
+        if (size > MAX_IMAGE_SIZE_MB) {
+            throw new ImageSizeExceededException(IMAGE_SIZE_EXCEED_MESSAGE);
+        }
+        this.size = size;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ImageFileSize imageFileSize = (ImageFileSize) o;
+        return size == imageFileSize.size;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(size);
+    }
+}
